requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
future==0.17.1
iso8601==0.1.12
more-itertools==4.3.0
paho-mqtt==1.4.0
pep8==1.7.1
pluggy==0.8.0
py==1.7.0
pyserial==3.4
pytest==4.0.1
PyYAML==3.13
six==1.12.0
